Refactor for multimodal simulation; documentation; bug fixes
