+++ bcipy/simulator/demo/README.md
@@ -1,0 +1,47 @@
+# Simulator Demo
+
+This module demonstrates how the simulator can be extended for various use cases.
+
+
+# Multimodal
+
+The `button_data_processor` and `button_press_model` are used to demonstrate a multimodal simulations using a button/switch as an example. To run a simulation with these inputs, you will need to perform the following steps:
+
+1. Ensure that the button signal model can be loaded or create a button signal model. To create a new one:
+
+```
+from pathlib import Path
+from bcipy.acquisition.datastream.mock.switch import switch_device
+from bcipy.io.save import save_model
+from bcipy.simulator.demo.button_press_model import ButtonPressModel
+from bcipy.signal.model.base_model import SignalModel, SignalModelMetadata
+
+dirname = "" # TODO: enter the directory
+model = ButtonPressModel()
+model.metadata = SignalModelMetadata(device_spec=switch_device(), evidence_type="BTN", transform=None)
+save_model(model, Path(dirname, "button_model.pkl"))
+```
+2. Ensure that the devices.json file has an entry for a switch
+
+```
+{
+    "name": "Switch",
+    "content_type": "MARKERS",
+    "channels": [
+        { "name": "Marker", "label": "Marker" }
+    ],
+    "sample_rate": 0.0,
+    "description": "Switch used for button press inputs",
+    "excluded_from_analysis": [],
+    "status": "active",
+    "static_offset": 0.0
+}
+```
+
+3. In your simulation parameters.json file, set the `acq_mode` parameter to 'EEG+MARKERS'.
+    - You may also want to set the `summarize_session` parameter to `true` to see how the evidences get combined during decision-making.
+4. Run a simulation.
+    - Select both the EEG and the Button models
+    - Use the InquirySampler
+
+Run with verbose mode and inspect the detailed run logs to ensure that the evidence is being sampled correctly.
